Revert "Move user pages to folder"
This reverts commit 3322785d9d446c440b8628fc1843495a1506a3bd.
