+++ frontend/src/ViewNurse.js
@@ -1,6 +1,6 @@
 import React from "react";
-import "../TableStyles.css";
-import Nav from "../Nav";
+import "./TableStyles.css";
+import Nav from "./Nav";
 
 class ViewNurse extends React.Component {
   constructor (props){
@@ -26,7 +26,7 @@ class ViewNurse extends React.Component {
                 window.location.href = "/edit_nurse";
                 }
             }>
-              <img src={require("../Icons/icons8-edit-64.png")} alt="edit" className="icon" />
+              <img src={require("./Icons/icons8-edit-64.png")} alt="edit" className="icon" />
             </button>
           </td>
         </tr>
